// src/shared/api/generated/models/telegramPersonalMessageStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type TelegramPersonalMessageStatus =
  (typeof TelegramPersonalMessageStatus)[keyof typeof TelegramPersonalMessageStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TelegramPersonalMessageStatus = {
  QUEUED: "QUEUED",
  SENDING: "SENDING",
  RETRY_WAIT: "RETRY_WAIT",
  SENT: "SENT",
  FAILED_PERMANENT: "FAILED_PERMANENT",
  OUTCOME_UNKNOWN: "OUTCOME_UNKNOWN",
  CANCELLED: "CANCELLED",
} as const;